requirements.txt
appdirs==1.4.0
backoff==1.4.0
google-api-python-client==1.6.2
google-auth==0.8.0
google-auth-httplib2==0.0.2
google-cloud-core==0.23.1
google-cloud-storage==0.23.0
googleapis-common-protos==1.5.2
httplib2==0.10.3
oauth2client==4.0.0
packaging==16.8
protobuf==3.2.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.1.10
requests==2.13.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
youtube-dl==2017.2.27
